requirements.txt
# --- Framework & Interface ---
streamlit==1.53.1
fastapi==0.128.0
uvicorn==0.40.0

# --- Cœur du Moteur RAG & IA ---
mistralai==1.10.1
faiss-cpu==1.13.2
sentence-transformers==5.2.0
torch==2.10.0
transformers==4.57.6

# --- Validation de Données & Modèles ---
pydantic==2.12.5
numpy==2.4.1
pandas==2.3.3

# --- Sécurité & Machine Learning (Guardrails) ---
scikit-learn==1.8.0

# --- Utilitaires & Environnement ---
python-dotenv==1.2.1
anyio==4.12.1
requests==2.32.5
httpx==0.28.1
tqdm==4.67.1

# --- Monitoring & Écologie ---
ecologits==0.9.2
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0

